Moved notify button to separate component.
